Ensure the relevant GPG release key is used
When having multiple GPG private keys on your keyring, this ensures the
one for the current project is picked. Requires that the key is stored
as "ProjectName release" in the key ring, where "ProjectName" is the
same as the last part of the GitHub URL, i.e.
git@github.com:OrganizationName/ProjectName

Signed-off-by: El RIDO <elrido@gmx.net>